fix code factor issues :art:
